refactor(section): Embedded를 활용하여 구간 일급 컬렉션 적용
