+++ src/main/java/nextstep/subway/domain/Sections.java
@@ -1,0 +1,21 @@
+package nextstep.subway.domain;
+
+import javax.persistence.CascadeType;
+import javax.persistence.Embeddable;
+import javax.persistence.OneToMany;
+import java.util.ArrayList;
+import java.util.List;
+
+@Embeddable
+public class Sections {
+    @OneToMany(mappedBy = "sections", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<Section> sections = new ArrayList<>();
+
+    public List<Station> getStations() {
+        return null;
+    }
+
+    public void add(Section section) {
+        sections.add(section);
+    }
+}
